Clean up dead code and style nits in the CIDR integration
- Remove unused library surface: get_cidr_mask(), set_cidr_mask(),
  irc_in6_CIDRMinIP() and the CIDR_SEARCH_ALL_MATCHES macros; nothing
  in ircd used them (the upward-walk case is covered by gliterIpMask).
- Use plain "res.h"-style includes instead of ../include/ paths; the
  build already passes -I$(top_srcdir)/include.
- Add the missing newline at the end of cidr_lookups.h.
- Parenthesize the family-dependent bit count in ircd_ntocidrmask();
  the old cast bound to the ternary condition, not the result.
- Drop the set-but-unused 'cidr' variable in make_gline() by moving
  the ircd_ntocidrmask() call inside the Debug() arguments, which
  compile away without DEBUGMODE.
- Convert stray // comments and commented-out DEBUG lines.

+++ ircd/ircd_string.c
@@ -465,7 +465,8 @@ const char* ircd_ntocidrmask(const struct irc_in_addr* in, const unsigned char b
 
   mask_str = ircd_ntoa(in);
   assert(strlen(mask_str) < 40);
-  snprintf(cidr, CIDR_LEN + 1, "%s/%u", mask_str, (unsigned int) irc_in_addr_is_ipv4(in) ? (bits - 96) : bits);
+  snprintf(cidr, CIDR_LEN + 1, "%s/%u", mask_str,
+           irc_in_addr_is_ipv4(in) ? (unsigned int)(bits - 96) : (unsigned int)bits);
   cidr[CIDR_LEN] = 0;
   return cidr;
 }
